+++ apps/client/src/app/components/portfolio-summary/portfolio-summary.component.ts
@@ -1,3 +1,4 @@
+import { NotificationService } from '@ghostfolio/client/core/notification/notification.service';
 import { getDateFnsLocale, getLocale } from '@ghostfolio/common/helper';
 import { PortfolioSummary, User } from '@ghostfolio/common/interfaces';
 import { translate } from '@ghostfolio/ui/i18n';
@@ -34,6 +35,8 @@ export class PortfolioSummaryComponent implements OnChanges {
   );
   public timeInMarket: string;
 
+  public constructor(private notificationService: NotificationService) {}
+
   public ngOnChanges() {
     if (this.summary) {
       if (this.summary.firstOrderDate) {
@@ -49,14 +52,15 @@ export class PortfolioSummaryComponent implements OnChanges {
   }
 
   public onEditEmergencyFund() {
-    const emergencyFundInput = prompt(
-      $localize`Please enter the amount of your emergency fund:`,
-      this.summary.emergencyFund?.total?.toString() ?? '0'
-    );
-    const emergencyFund = parseFloat(emergencyFundInput?.trim());
+    this.notificationService.prompt({
+      confirmFn: (value) => {
+        const emergencyFund = parseFloat(value.trim()) || 0;
 
-    if (emergencyFund >= 0) {
-      this.emergencyFundChanged.emit(emergencyFund);
-    }
+        this.emergencyFundChanged.emit(emergencyFund);
+      },
+      confirmLabel: $localize`Save`,
+      defaultValue: this.summary.emergencyFund?.total?.toString() ?? '0',
+      title: $localize`Please set the amount of your emergency fund.`
+    });
   }
 }